requirements.txt
dj-database-url==0.5.0
Django==2.0.4
django-model-utils==3.1.1
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.4
whitenoise==3.3.1
